Everywhere: Correct and convert `netMode` checks
After a realization and conversation with TML, it is bogus to check the
`netMode` during `Load` because the life cycle of mod loading has
already completed. The only thing that makes sense is checking for
server, which has caused most of our code to accidentally work.

This changes many `netMode` checks into `dedServ` checks instead, and
makes `AdventureInventory` only autoload on the client.