PointsManager+AdventureNpc: Fix tracking of Eater of Worlds
The Eater of Worlds can be damaged at multiple segments (3 different
NPCs), and the last segment killed is marked as the boss. The segment
marked as the boss may additionally be killed by AI instead of by the
player.

To resolve this, propagate damage done to ANY segments to ALL segments.
This ensures that no matter which segment dies, it is always attributed
accurately.

Additionally, we will mark ALL segments as downed for the world, in the
event different segments marked as the boss die.