Adventure{Player,Projectile,Item,Config}: Add PvP damage modifications
This configuration allows you to change the damage multiplier dealt by
specific items for PvP.

To help in this with projectiles, we need to store the `IEntitySource`
of every projectile with itself, and modify
`PlayerDeathReason.ByProjectile` to include the `Item` if the entity
source can supply it. This way, the source item can be referenced
instead of just the projectile type when inside `ModifyHurt`.