AdventurePlayer+CombatManager: Introduce damage falloff & configuration
For configured items and projectiles, the damage falloff multiplier can
be adjusted. This is calculated by means of exponential decay.

This also includes a fix for `ModPlayer.ModifyHurt` in PvP. Previously,
calling `Player.Hurt` for players during PvP would _not_ generate a
direct hit, and all clients would invoke `ModPlayer.ModifyHurt` -- which
is unexpected and invalid.

By loosening some restraints in the networking of hurt packets in
`Player.Hurt`, itself can now properly network direct hits as a result
of PvP.

An additional method is stubbed out, as it is legacy, internal pvp-only
netcode that supported what we now removed.