Scoreline: Correct team order and don't show empty teams
This slightly changes the status-quo, where we prioritized teams
appearing on the right side, as teams now begin to appear from the left.
Doesn't actually matter and makes it easier to get the order correct.

This also cleans up the code quite a bit, removing some magic numbers
that should have been using the constants.