AdventureConfig: Add equality and hash code overrides for `Statistics`
Without this, tModLoader would _never_ consdier two instances of this to
be equal, meaning it would always require reloads any moment it
questioned if it needed one.

A lot of this stems from the unfortunately requirements already
complained about in comments above -- we are bending backwards for
tModLoader's config and UI to function in any meaningful way.