AdventureNpc: Allow Wall of Flesh to be summoned using either ritual
Because our town NPCs are invincible, the feature of being able to drop
the Guide in lava to summon the Wall of Flesh doesn't operate correctly.
This fixes that, allowing the Guide to _only_ die when actually spawning
Wall of Flesh.

This also removes the requirement for the Guide to be alive when
dropping a Guide Voodoo Doll into lava.